# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shared SILA ME foundation API for Cloud and Mobile
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/cloud/login:
    post:
      operationId: cloudLogin
      tags: [auth]
      summary: Start a Cloud session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated Cloud account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          $ref: "#/components/responses/AuthenticationError"
  /auth/cloud/session:
    get:
      operationId: getCloudSession
      tags: [auth]
      summary: Restore a Cloud session
      responses:
        "200":
          description: Authenticated Cloud account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          $ref: "#/components/responses/AuthenticationError"
...
  /auth/mobile/login:
    post:
      operationId: mobileLogin
      tags: [auth]
      summary: Start a Mobile session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated Mobile account and bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileAuthSession"
        "401":
          $ref: "#/components/responses/AuthenticationError"
  /auth/mobile/logout:
    post:
      operationId: mobileLogout
      tags: [auth]
      summary: Revoke the current Mobile session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session revoked
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: sila_me_session
  responses:
    AuthenticationError:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
    CurrentUser:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
        application:
          type: string
          enum: [CLOUD, MOBILE]
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
      required: [id, displayName, email, application, status]
    AuthSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/CurrentUser"
      required: [user]
    MobileAuthSession:
      allOf:
        - $ref: "#/components/schemas/AuthSession"
        - type: object
          properties:
            token:
              type: string
          required: [token]
    ApiError:
      description: Structured operational error returned to preserve recoverable validation and duplicate details.
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        missingFields:
          description: Input fields that must be completed before the invoice can be saved.
          type: array
          items: { type: string }
          nullable: true
        existingInvoiceId:
          description: Existing invoice that caused a duplicate warning, when available.
          type: string
          format: uuid
          nullable: true
        duplicateType:
          description: Duplicate classification, such as PROBABLE.
          type: string
          nullable: true
      required: [code, message]
...